inject DroidKaigiService and GithubService
