+++ app/src/main/java/io/github/droidkaigi/confsched2017/api/DroidKaigiClient.java
@@ -1,10 +1,5 @@
 package io.github.droidkaigi.confsched2017.api;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-
-import com.jakewharton.retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory;
-
 import android.support.annotation.NonNull;
 
 import java.util.List;
@@ -17,9 +12,6 @@ import io.github.droidkaigi.confsched2017.api.service.GithubService;
 import io.github.droidkaigi.confsched2017.model.Contributor;
 import io.github.droidkaigi.confsched2017.model.Session;
 import io.reactivex.Single;
-import okhttp3.OkHttpClient;
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
 
 @Singleton
 public class DroidKaigiClient {
@@ -33,25 +25,9 @@ public class DroidKaigiClient {
     private static final int MAX_PER_PAGE = 100;
 
     @Inject
-    public DroidKaigiClient(OkHttpClient client) {
-        Retrofit droidkaigiRetrofit = new Retrofit.Builder()
-                .client(client)
-                .baseUrl("https://droidkaigi.github.io")
-                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-                .addConverterFactory(GsonConverterFactory.create(createGson()))
-                .build();
-        droidKaigiService = droidkaigiRetrofit.create(DroidKaigiService.class);
-
-        Retrofit githubRetrofit = new Retrofit.Builder().client(client)
-                .baseUrl("https://api.github.com")
-                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-                .addConverterFactory(GsonConverterFactory.create(createGson()))
-                .build();
-        githubService = githubRetrofit.create(GithubService.class);
-    }
-
-    private static Gson createGson() {
-        return new GsonBuilder().setDateFormat("yyyy/MM/dd HH:mm:ss").create();
+    public DroidKaigiClient(DroidKaigiService droidKaigiService, GithubService githubService) {
+        this.droidKaigiService = droidKaigiService;
+        this.githubService = githubService;
     }
 
     public Single<List<Session>> getSessions(@NonNull String languageId) {
